Touch: Move tests to it's own file and create a struct to avoid calls to ArgMatches
